backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.32.5
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.1
# pinecone-client==3.0.0  # Optional - only needed for context memory
